requirements.txt
numpy
scipy
matplotlib
tme
flax
optax
jaxopt
pytest
einops
numpyro
scikit-image
